==> Step 2 - Data Preprocessing.ipynb <==
# Data Preprocessing

Install Library

pip install swifter

pip install openpyxl

pip install Sastrawi

### Import Packages and Dataset

import pickle
import pandas as pd
import numpy as np
import re
import swifter
import matplotlib.pyplot as plt
import seaborn as sns
import time
import csv
from nltk.tag import CRFTagger
from nltk.tokenize import word_tokenize
from textblob import TextBlob
from random import random
from openpyxl import Workbook

sns.set_style("whitegrid")

from sklearn.model_selection import train_test_split

from Sastrawi.StopWordRemover.StopWordRemoverFactory import StopWordRemoverFactory
from Sastrawi.Stemmer.StemmerFactory import StemmerFactory

import warnings
warnings.filterwarnings("ignore")

# membuat akses ke google drive untuk load data

from google.colab import drive
drive.mount('/content/drive', force_remount=True)

full_data_train = pd.read_csv('/content/drive/MyDrive/TA_MEIRA REYNITA PUTRI_1301164695/Code/Data/data_train_70_30_1.csv', encoding = 'unicode_escape', sep =',')
full_data_test = pd.read_csv('/content/drive/MyDrive/TA_MEIRA REYNITA PUTRI_1301164695/Code/Data/data_test_70_30_1.csv', encoding = 'unicode_escape', sep = ',')

full_data_train = full_data_train.reset_index(drop = True)
full_data_train.sample(5)

full_data_test = full_data_test.reset_index(drop = True)
full_data_test.sample(5)

### Text Cleaning

1. Menghilangkan simbol khusus

def cleaning_data(data):
    data['Argumen_Cleaned_1'] = data['Argumen'].str.replace("\r", " ")
    data['Argumen_Cleaned_1'] = data['Argumen_Cleaned_1'].str.replace("\t", " ")
    data['Argumen_Cleaned_1'] = data['Argumen_Cleaned_1'].str.replace("\n", " ")
    return data

full_data_train = cleaning_data(full_data_train)
full_data_test = cleaning_data(full_data_test)

2. Ubah menjadi lowercase

def lowercase(data):
    data['Argumen_Cleaned_2'] = data['Argumen_Cleaned_1'].str.lower()
    return data

full_data_train = lowercase(full_data_train)
full_data_test = lowercase(full_data_test)

3. Menghilangkan emoji, simbol, dan tanda baca

RE_EMOJI = re.compile('[\U00010000-\U0010ffff]', flags=re.UNICODE)

def strip_emoji(text):
    return RE_EMOJI.sub(r'', text)

def remove_emoji(data):
    data['Argumen_Cleaned_3'] = data['Argumen_Cleaned_2']
    data['Argumen_Cleaned_3'] = data['Argumen_Cleaned_3'].map(lambda x: strip_emoji(x))
    return data

full_data_train = remove_emoji(full_data_train)
full_data_test = remove_emoji(full_data_test)

def remove_punct(data):
    data['Argumen_Cleaned_4'] = data['Argumen_Cleaned_3']
    data['Argumen_Cleaned_4'] = data['Argumen_Cleaned_4'].str.replace("[^a-zA-Z0-9]+", " ")
    data['Argumen_Cleaned_4'] = data['Argumen_Cleaned_4'].str.replace('\u2026', '')
    data['Argumen_Cleaned_4'] = data['Argumen_Cleaned_4'].str.replace("  ", " ")
    data['Argumen_Cleaned_4'] = data['Argumen_Cleaned_4'].str.replace("   ", " ")
    data['Argumen_Cleaned_4'] = data['Argumen_Cleaned_4'].str.replace("’s", "")
    data['Argumen_Cleaned_4'] = data['Argumen_Cleaned_4'].str.replace("'s", "")
    data['Argumen_Cleaned_4'] = data['Argumen_Cleaned_4'].str.replace("n't"," not")
    data['Argumen_Cleaned_4'] = data['Argumen_Cleaned_4'].str.replace("'re"," are")
    data['Argumen_Cleaned_4'] = data['Argumen_Cleaned_4'].str.replace("'tn"," nt")
    return data

full_data_train = remove_punct(full_data_train)
full_data_test = remove_punct(full_data_test)

4. Menghilangkan Angka

def removeNumber(data):
    data['Argumen_Cleaned_4'] = data['Argumen_Cleaned_4'].str.replace("[^a-zA-Z0-9]+", " ")
    return data

full_data_train = removeNumber(full_data_train)
full_data_test = removeNumber(full_data_test)

### Normalisasi Kata 

norm = pd.read_csv('/content/drive/MyDrive/TA_MEIRA REYNITA PUTRI_1301164695/Code/Data/Normalisasi_Kata.csv', encoding = 'unicode_escape')
norm.head()

kata_tak_baku = norm['tak_baku'].values.tolist()
kata_baku = norm['baku'].values.tolist()

def normalisasi_kata(kata, baku, tak_baku):
    kata = kata.split(" ")
    for j in range(len(kata)):
        if kata[j] in tak_baku:
            a = tak_baku.index(kata[j])
            kata[j] = baku[a]
    return ' '.join(map(str,kata))
    
def normalisasi_data(data, baku, tak_baku):
    data['Argumen_Normed'] = data['Argumen_Cleaned_4']
    data['Argumen_Normed'] = data['Argumen_Normed'].map(lambda x: normalisasi_kata(x,baku,tak_baku))
    return data

full_data_train = normalisasi_data(full_data_train, kata_baku, kata_tak_baku)
full_data_test = normalisasi_data(full_data_test, kata_baku, kata_tak_baku)

### Menghilangkan Stopword

factory_stop = StopWordRemoverFactory()
stopword = factory_stop.get_stop_words()

unwanted_list = ['namun', 'begitu', 'ok', 'oke', 'tidak', 'tetapi', 'tapi', 'amat', 'seharusnya', 'sebetulnya', 'setidaknya', 'tentu', 'bagaimanapun', 'daripada', 'sedangkan', 'sementara', 'agak', 'yang', 'dan', 'di', 'dari', 'nomor', 'tahun', 'no']
final_stopword = []

for element in stopword:
    if element not in unwanted_list:
        final_stopword.append(element)

def remove_stop(data):
    data['Argumen_Stopword'] = data['Argumen_Normed']

    for stop in final_stopword:
        re_stop = r" \b" + stop + r"\b "
        data['Argumen_Stopword'] = data['Argumen_Stopword'].str.replace(re_stop, ' ')
        # print(re_stop)
    return data

full_data_train = remove_stop(full_data_train)
full_data_test = remove_stop(full_data_test)

### Stemming Kata

factory_stem = StemmerFactory()
stemmer = factory_stem.create_stemmer()

def stemming_kata(kata):
    kata = kata.split(" ")
    for j in range(len(kata)):
        stem = [stemmer.stem(y) for y in kata]
    return ' '.join(map(str,stem))

def stemming_data(data):
    data['Argumen_Stemmed'] = data['Argumen_Stopword']
    data['Argumen_Stemmed'] = data['Argumen_Stemmed'].swifter.apply(lambda x: stemming_kata(x))
    return data

full_data_train = stemming_data(full_data_train)
full_data_test = stemming_data(full_data_test)

### Fixing Data

def fixData(data):
      data_fix = data.drop(['Argumen_Cleaned_1', 'Argumen_Cleaned_2', 'Argumen_Cleaned_3', 'Argumen_Cleaned_4', 'Argumen_Normed', 'Argumen_Stopword'], axis=1)
      data_fix = data_fix.rename(columns={'Argumen_Stemmed': 'Argumen_Fixed'})
      return data_fix

full_data_train = fixData(full_data_train)
full_data_test = fixData(full_data_test)

### Save Data

full_data_train.to_csv('/content/drive/MyDrive/TA_MEIRA REYNITA PUTRI_1301164695/Code/Data/full_data_train_1.csv')
full_data_test.to_csv('/content/drive/MyDrive/TA_MEIRA REYNITA PUTRI_1301164695/Code/Data/full_data_test_1.csv')
