==> Step 3 - Feature Extraction.ipynb <==
## Feature Extraction

### Import Packages and Dataset

import pandas as pd

from sklearn.feature_extraction.text import TfidfVectorizer, CountVectorizer

import warnings
warnings.filterwarnings("ignore")

# membuat akses ke google drive untuk load data

from google.colab import drive
drive.mount('/content/drive', force_remount=True)

full_data_train = pd.read_csv('/content/drive/MyDrive/TA_MEIRA REYNITA PUTRI_1301164695/Code/Data/full_data_train.csv')
full_data_test = pd.read_csv('/content/drive/MyDrive/TA_MEIRA REYNITA PUTRI_1301164695/Code/Data/full_data_test.csv')

### Feature Target Split

X_train = full_data_train['Argumen_Fixed']
X_test = full_data_test['Argumen_Fixed']

y_train = full_data_train.drop(['Unnamed: 0', 'Argumen', 'Argumen_Fixed'], axis=1)
y_test = full_data_test.drop(['Unnamed: 0', 'Argumen', 'Argumen_Fixed'], axis=1)

### TF-IDF

def tfidf(X_train,X_test,y_train,y_test,target,max_features = 300):
    ngram_range = (1,2)
    min_df = 10
    max_df = 1.
    tfidf = TfidfVectorizer(encoding='utf-8',
                        ngram_range=ngram_range,
                        stop_words=None,
                        lowercase=False,
                        max_df=max_df,
                        min_df=min_df,
                        max_features=max_features,
                        norm='l2')
    features_train = tfidf.fit_transform(X_train).toarray()
    labels_train = y_train.values
    features_test = tfidf.transform(X_test).toarray()
    labels_test = y_test.values
    features_all = tfidf.get_feature_names()
    print(features_all)
    data_train_tfidf = pd.DataFrame(data=features_train,columns=features_all)
    data_train_tfidf = data_train_tfidf.join(pd.DataFrame(data=labels_train,columns=target))
    data_test_tfidf = pd.DataFrame(data=features_test,columns=features_all)
    data_test_tfidf = data_test_tfidf.join(pd.DataFrame(data=labels_test,columns=target))
    return data_train_tfidf,data_test_tfidf

target_1 = full_data_test.drop(['Unnamed: 0','Argumen','Argumen_Fixed'], axis=1).columns

train_data_tfidf, test_data_tfidf = tfidf(X_train, X_test, y_train, y_test, target_1,300)

print(train_data_tfidf.shape)
print(test_data_tfidf.shape)

train_data_tfidf

test_data_tfidf

train_data_tfidf.to_csv("/content/drive/MyDrive/TA_MEIRA REYNITA PUTRI_1301164695/Code/Data/data_train_tfidf.csv", index=False)
test_data_tfidf.to_csv("/content/drive/MyDrive/TA_MEIRA REYNITA PUTRI_1301164695/Code/Data/data_test_tfidf.csv", index=False)

### TF-RF Using Vectorizer

def tfrf(X_train,X_test,y_train,y_test,target,max_features = 300):
    ngram_range = (1,2)
    min_df = 10
    max_df = 1.
    tfrf = CountVectorizer(encoding='utf-8',
                        ngram_range=ngram_range,
                        stop_words=None,
                        lowercase=False,
                        max_df=max_df,
                        min_df=min_df,
                        max_features=max_features,)
    features_train = tfrf.fit_transform(X_train).toarray()
    labels_train = y_train.values
    features_test = tfrf.transform(X_test).toarray()
    labels_test = y_test.values
    features_all = tfrf.get_feature_names()
    print(features_all)
    data_train_tfrf = pd.DataFrame(data=features_train,columns=features_all)
    data_train_tfrf = data_train_tfrf.join(pd.DataFrame(data=labels_train,columns=target))
    data_test_tfrf = pd.DataFrame(data=features_test,columns=features_all)
    data_test_tfrf = data_test_tfrf.join(pd.DataFrame(data=labels_test,columns=target))
    return data_train_tfrf,data_test_tfrf

target_2 = full_data_test.drop(['Unnamed: 0','Argumen','Argumen_Fixed'], axis=1).columns

train_data_tfrf, test_data_tfrf = tfrf(X_train, X_test, y_train, y_test, target_2,300)

print(train_data_tfrf.shape)
print(test_data_tfrf.shape)

train_data_tfrf

test_data_tfrf

train_data_tfrf.to_csv("/content/drive/MyDrive/TA_MEIRA REYNITA PUTRI_1301164695/Code/Data/data_train_tfrf.csv", index=False)
test_data_tfrf.to_csv("/content/drive/MyDrive/TA_MEIRA REYNITA PUTRI_1301164695/Code/Data/data_test_tfrf.csv", index=False)
