==> Step 4 - Training Classifier.ipynb <==
## Training Classifier

### Import Packages and Dataset

import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns

sns.set_style("whitegrid")

from sklearn import svm
from sklearn.model_selection import RandomizedSearchCV, GridSearchCV, ShuffleSplit
from sklearn.metrics import f1_score, recall_score, precision_score, confusion_matrix, accuracy_score, classification_report

import warnings
warnings.filterwarnings("ignore")

from google.colab import drive
drive.mount('/content/drive')

data_train_tfidf = pd.read_csv("/content/drive/MyDrive/TA_MEIRA REYNITA PUTRI_1301164695/Code/Data/data_train_tfidf.csv")
data_test_tfidf = pd.read_csv("/content/drive/MyDrive/TA_MEIRA REYNITA PUTRI_1301164695/Code/Data//data_test_tfidf.csv")

full_data_train = pd.read_csv('/content/drive/MyDrive/TA_MEIRA REYNITA PUTRI_1301164695/Code/Data/full_data_train.csv')
full_data_test = pd.read_csv('/content/drive/MyDrive/TA_MEIRA REYNITA PUTRI_1301164695/Code/Data/full_data_test.csv')

X_train = data_train_tfidf.drop(['Kelas'], axis=1)
y_train = data_train_tfidf[['Kelas']]
X_test = data_test_tfidf.drop(['Kelas'], axis=1)
y_test = data_test_tfidf[['Kelas']]

### Support Vector Machine Classifier

#classifier = svm.SVC() 
classifier = svm.SVC(kernel='rbf', C = 100, gamma= 100) #degree = 3

classifier.fit(X_train, y_train)

predict = classifier.predict(X_test)

predict

print("F1 Score         : ", f1_score(y_test, predict,
                                           pos_label='positive',
                                           average='weighted'))
print("Akurasi Score    : ", accuracy_score(y_test, predict))
print("Precision Score  : ", precision_score(y_test, predict, 
                                           pos_label='positive',
                                           average='weighted'))
print("Recall Score     : ", recall_score(y_test, predict, 
                                           pos_label='positive',
                                           average='weighted'))

classifier.fit(X_train, y_train)
baseline_classifier = classifier.predict(X_test)
accuracy_score(y_test, baseline_classifier)

print("Classification report")
print(classification_report(y_test, baseline_classifier))

##tfrf

table = pd.DataFrame({
    'Argumen'             : full_data_test['Argumen_Fixed'],
    'Kelas'               : full_data_test['Kelas'],
    'Kelas Prediksi Test' : baseline_classifier,
})

table.sample(10)

table.to_excel("/content/drive/MyDrive/TA_MEIRA REYNITA PUTRI_1301164695/Code/Data/result_svm_1.xlsx", index=False)

data_train_tfrf = pd.read_csv("/content/drive/MyDrive/TA_MEIRA REYNITA PUTRI_1301164695/Code/Data/data_train_tfrf.csv")
data_test_tfrf = pd.read_csv("/content/drive/MyDrive/TA_MEIRA REYNITA PUTRI_1301164695/Code/Data//data_test_tfrf.csv")

X_train = data_train_tfrf.drop(['Kelas'], axis=1)
y_train = data_train_tfrf[['Kelas']]
X_test = data_test_tfrf.drop(['Kelas'], axis=1)
y_test = data_test_tfrf[['Kelas']]

##TFRF

#classifier = svm.SVC() 
classifier = svm.SVC(kernel='rbf', C = 100, gamma = 100 ) #degree = 3

classifier.fit(X_train, y_train)

predict = classifier.predict(X_test)

predict

print("F1 Score         : ", f1_score(y_test, predict,
                                           pos_label='positive',
                                           average='weighted'))
print("Precision Score  : ", accuracy_score(y_test, predict))
print("Precision Score  : ", precision_score(y_test, predict, 
                                           pos_label='positive',
                                           average='weighted'))
print("Recall Score     : ", recall_score(y_test, predict, 
                                           pos_label='positive',
                                           average='weighted'))

classifier.fit(X_train, y_train)
baseline_classifier = classifier.predict(X_test)
accuracy_score(y_test, baseline_classifier)

print("Classification report")
print(classification_report(y_test, baseline_classifier))

table = pd.DataFrame({
    'Argumen'             : full_data_test['Argumen_Fixed'],
    'Kelas'               : full_data_test['Kelas'],
    'Kelas Prediksi Test' : baseline_classifier,
})

table.sample(10)

table.to_excel("/content/drive/MyDrive/TA_MEIRA REYNITA PUTRI_1301164695/Code/Data/result_svm_2.xlsx", index=False)
